+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -572,5 +572,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-24T05:41:46.101Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-24T06:51:10.404Z")
     r"""The date and time the resource was last updated"""